Refactor mfatfm package layout
