+++ src/mfatfm/__init__.py
@@ -1,0 +1,15 @@
+"""Public interface for the mfatfm package."""
+
+from .connection import AsyncSFTPManager
+from .fileops import FileEntry, load_local_directory, normalize_local_path
+from .ui import FileManagerWindow, SFTPProgressDialog, launch_file_manager_window
+
+__all__ = [
+    "AsyncSFTPManager",
+    "FileEntry",
+    "FileManagerWindow",
+    "SFTPProgressDialog",
+    "launch_file_manager_window",
+    "load_local_directory",
+    "normalize_local_path",
+]
